Create a new commit from the changes.
This takes whatever you've added to the index, and packages the change up into
a commit. The commit includes all the file changes, including the filee paths,
the date of the commit, as well as your name and email address.

It also includes the hash of the commit (or commits) that came before it. This
is how git knows how to efficiently figure out when a particular line of code
was changed. Every commit has a reference back to its 'parent.'