Merge a named commit into where you are
The git merge [branchname] command will merge the changes in the named branch
into your current branch. It does this by 'replaying' all the commits that
happened in the named branch over again, taking into account when they were
made.

Once it has replayed all the commits from the named branch, it will create a
new commit that has TWO parents; first, the original parent from where you
currently are in the repository, and second, the commit you're merging in.

This creates a special kind of commit: A merge commit.